requirements.txt
ccxt
python-dotenv
requests
pandas
numpy
